+++ pakyow-js/__tests__/features/support/helpers/sleep.js
@@ -1,0 +1,5 @@
+export default function(ms) {
+  return new Promise(resolve => {
+    setTimeout(resolve, ms);
+  });
+}