+++ pakyow-presenter/lib/pakyow/presenter/renderer/behavior/place_in_mode.rb
@@ -6,6 +6,7 @@ module Pakyow
   module Presenter
     class Renderer
       module Behavior
+        # @api private
         module PlaceInMode
           extend Support::Extension