+++ pakyow-assets/CHANGELOG.md
@@ -1,5 +1,7 @@
 # UNRELEASED
 
+  * [added] Devtools component as an autoloaded pack in development and prototype modes
+
 # 1.0
 
   * Hello, Web