+++ pakyow-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # UNRELEASED
 
+  * [added] Environment action for triggering restarts at an http endpoint
+    * Only available when running development and prototype environments
+  * [added] Ability to restart or respawn into a particular environment by writing to the tmp file
+  * [added] Explicit names to environment actions for attaching new behavior before/after
+  * [changed] Reuse the same port for respawns, just like we do restarts
+
 # 1.0.1
 
   * Rename "navigable" to "navigator" in the generated app