+++ pakyow-support/CHANGELOG.md
@@ -1,5 +1,9 @@
 # UNRELEASED
 
+  * [changed] `Config` now supports multiple default config blocks per environnment
+  * [changed] `Config#defaults` now supports passing multiple environments
+  * [fixed] Bug that occurs if a named action shares a name with a method on the pipelined object
+
 # 1.0
 
   * Hello, Web